Add v4 docs
Bourbon v5 runs on a whole new documentation system (SassDoc) and for
the first time, the documentation site will be versioned going forward.
This breaks away from our current method, which is a Jekyll site where
the docs are manually written and maintained.

Since v5 also carries a lot of major, breaking changes, we want to make
sure our current v4 docs are still publicly available, for those who
choose to stay with that version.

This commit adds an "archived" version of the existing v4 docs page,
along with its CSS. It uses a custom layout, so it's isolated from all
of the new v5 docs.

Closes #31