Denote current page links using aria-current
The middleman-aria_current gem provides a `current_link_to` helper,
which wraps the built-in `link_to` helper. It checks the URL of the
current visited page and outputs an `aria-current` attribute if it
matches the URL in the link.

If you're new to `aria-current`, Léonie Watson wrote a wonderful article
detailing its usage: http://tink.uk/using-the-aria-current-attribute/

The spec for `aria-current` can be found here: https://www.w3.org/TR/wai-aria-1.1/#aria-current